v0.0.3- Editing email fields Model User
